finishes websockets messaging functions

+++ frontend/components/messages/message_body.jsx
@@ -4,17 +4,26 @@ function MessageBody(props) {
   const [body, setBody] = useState(props.message.body)
   const [edit, setEdit] = useState(false)
 
+
+  const messageOptions = () => {
+    if (props.currentUserId === props.message.user.id) {
+      return (
+        <div className="message-button-container">
+          <div onClick={() => setEdit(!edit)}>edit</div>
+          <div>options</div>
+          <div onClick={()=> props.deleteMessage(props.message.id)}>Delete</div>
+        </div>
+      )
+    }
+  }
+
   const messageBody = () => {
     return (
       <div>
         <div>
           {body}
         </div>
-        <div className="message-button-container">
-          <div onClick={() => setEdit(!edit)}>edit</div>
-          <div>options</div>
-          <div onClick={()=> props.deleteMessage(props.message.id)}>Delete</div>
-        </div>
+        {messageOptions()}
       </div>
     )
   }
